requirements.txt
python-dotenv
requests
trafilatura
playwright
bs4
waybackpy